+++ src/contents/pomodoro.tsx
@@ -22,7 +22,7 @@ export default function PomodoroContent(){
                     <h2 className={cn("text-muted-foreground text-center",runtime.phase==="idle" ? "text-sm md:text-base" : "text-lg md:text-xl lg:text-2xl")}>{
                          runtime.phase!=="idle" ? t(`phase.${runtime.phase}`) : t("tip")
                     }</h2>
-                    <h2 className="text-6xl md:text-7xl">{minutes}:{seconds}</h2>
+                    <h2 className="text-4xl xs:text-5xl sm:text-6xl md:text-7xl">{minutes}:{seconds}</h2>
                     <Progress value={(1 - runtime.remaining / runtime.total) * 100}/>
                     <div className="flex justify-center items-center flex-wrap gap-2 w-full">
                          <Button className="flex-1" onClick={() => runtime.phase === "idle" ? start() : runtime.isPaused ? resume() : pause()}>