+++ src/pages/exercise.tsx
@@ -1,10 +1,15 @@
 import BreathingExerciseProvider from "@/context/breathing-exercise";
-import BreathingExerciseContent from "@/contents/breathing-exercise";
+import BreathingExerciseLoader from "@/loaders/breathing-exercise";
+import { lazy, Suspense } from "react";
+
+const BreathingExerciseContent = lazy(()=>import("@/contents/breathing-exercise"));
 
 export default function BreathingExercisePage(){
      return (
           <BreathingExerciseProvider>
-               <BreathingExerciseContent/>
+               <Suspense fallback={<BreathingExerciseLoader/>}>
+                    <BreathingExerciseContent/>
+               </Suspense>
           </BreathingExerciseProvider>
      )
 }