+++ src/pages/pomodoro.tsx
@@ -1,10 +1,15 @@
-import PomodoroContent from "@/contents/pomodoro";
 import { PomodoroProvider } from "@/context/pomodoro";
+import PomodoroLoader from "@/loaders/pomodoro";
+import { lazy, Suspense } from "react";
+
+const PomodoroContent = lazy(()=>import("@/contents/pomodoro"));
 
 export default function PomodoroPage(){
      return (
           <PomodoroProvider>
-               <PomodoroContent/>
+               <Suspense fallback={<PomodoroLoader/>}>
+                    <PomodoroContent/>
+               </Suspense>
           </PomodoroProvider>
      )
 }